content/canvas/test/test_2d.line.cap.invalid.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.line.cap.invalid - bug 401788</title>
<!-- Testing: Setting lineCap to invalid values is ignored -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

ctx.lineCap = 'butt'
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'invalid';
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'ROUND';
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'round\0';
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'round ';
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = "";
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'bevel';
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
